+++ module/customize.sh
@@ -23,6 +23,8 @@ set_perm_recursive "$MODPATH/data" root root 0700 0600
 set_perm_recursive "$MODPATH/certs" root root 0755 0644
 set_perm_recursive "$MODPATH/certs/custom" root root 0700 0600
 set_perm_recursive "$MODPATH/certs/sources" root root 0700 0600
+[ -d "$MODPATH/system/etc/security/cacerts" ] && \
+  set_perm_recursive "$MODPATH/system/etc/security/cacerts" root root 0755 0644
 [ -d "$MODPATH/webroot" ] && set_perm_recursive "$MODPATH/webroot" root root 0755 0644
 for s in post-fs-data.sh service.sh action.sh uninstall.sh customize.sh; do
   [ -f "$MODPATH/$s" ] && set_perm "$MODPATH/$s" root root 0755